Move driver loading into a function
- Will be re-used by the writer driver
- Also deleting obsolete documentation
